Refactor the representations for requests and responses, while implementing `UVerb` headers (#22)
While trying to implement a solution to #13, I realised that we can't use the same `Body` type to represent requests and responses, mostly as a result of the way that Servant supports multiple alternate responses (via `UVerb` and `Union`).

For requests, multiple `ReqBody`s chained together will make Servant try to parse the request body as _all_ of the types; moreover, any `Header`s in the request will be parsed indpendently of the body.
However for responses, multiple possible responses are represented using `UVerb` (as opposed to `Verb`), which bundles together response headers and response bodies.

Therefore we need to change the representation.
The one common abstraction we can use in both situations is that we want to express whether there is 0, 1, or multiple interpretations. That's what the `Some` type does. But we want `Some` to contain different types: for requests we only care about the `TypeRep` of the request type, whereas for responses we care about both the `TypeRep` and the `Header`s, which we bundle together in the `Response` type.