Move `.golden` inside `servant-routes-golden`
Turns out there's a difference between `cabal run servant-routes-golden-test`,
which runs the test suite with the working directory set to the project
root; and `cabal test servant-routes-golden`, which runs it with the
working directory set to `servant-routes-golden`. The CICD does the
latter, so the `.golden` directory needs to be in that directory too
(otherwise the tests pass trivially).